Extract pure functions to new files
Improve code organization, reuse, and testability by extracting pure
JavaScript functions to new files outside of the React structure. I've
also added tests for the shuffle function, which is currently failing.

The test is failing because shuffle currently *never* keeps an item
in it's original index. I.e. an item at index 0 in the input array
will never be at index 0 in the shuffled array. The counts variabe
after one test run were:

```json
{
    a: [0, 5102, 4898],
    b: [4898, 0, 5102],
    c: [5102, 4898, 0]
}
```

I will fix that in a subsequent commit.
